+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.2.2 (2020-05-29)
+==================
+
+- added -i, --info argument to return information about current version
+  as json
+- print warnings to stderr
+
 0.2.1 (2020-05-26)
 ==================